adding controllers

+++ Social_media_platfom/src/main/java/com/example/social/serviceImpl/PostServiceImpl.java
@@ -8,10 +8,10 @@ import java.util.Optional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import com.example.repository.PostRepo;
-import com.example.services.PostService;
 import com.example.social.Exception.PostException;
 import com.example.social.model.Post;
+import com.example.social.repository.PostRepo;
+import com.example.social.services.PostService;
 
 @Service
 public class PostServiceImpl implements PostService{
@@ -28,8 +28,16 @@ public class PostServiceImpl implements PostService{
         return postRepository.findById(id);
     }
     
-    public void deletePostById(Long id) throws PostException{
-        postRepository.deleteById(id);
+    public boolean deletePostById(Long id) throws PostException{
+    	Optional<Post> post = postRepository.findById(id);
+    	if(post.isPresent()) {
+    		postRepository.deleteById(id);
+    		return true;
+    	}else {
+    		return false;
+    		
+    	}
+        
     }
     
     public void updatePostContentById(Long id, String content) throws PostException{
@@ -43,22 +51,24 @@ public class PostServiceImpl implements PostService{
         
     }
     
-    public void incrementLikesById(Long id) throws PostException{
+    public Post incrementLikesById(Long id) throws PostException{
     	Optional<Post> post = postRepository.findById(id);
     	if(post.isPresent()) {
     		Post p = post.get();
     		p.setLikes(p.getLikes()+1);;
+    		return p;
     	}else {
     		throw new PostException("Id not available");
     	}
     }
     
-    public void decrementLikesById(Long id) throws PostException{
+    public Post decrementLikesById(Long id) throws PostException{
     	Optional<Post> post = postRepository.findById(id);
     	if(post.isPresent()) {
     		Post p = post.get();
     		if(p.getLikes()>0) {
     			p.setLikes(p.getLikes()-1);
+    			return p;
     		}
     		else {
     			throw new PostException("Likes can not be negetive");
@@ -68,7 +78,7 @@ public class PostServiceImpl implements PostService{
     	}
     }
     
-    public List<Post> getTopLikedPosts(int limit) throws PostException{
+    public List<Post> getTopLikedPosts() throws PostException{
     	List<Post> posts = postRepository.findAll();
     	if(posts.size()>=5) {
     		posts.sort((p1,p2)->{
